SImplified log forwarding by using webhook lib
