+++ application/build.gradle
@@ -50,6 +50,8 @@ dependencies {
     implementation 'org.apache.logging.log4j:log4j-core:2.19.0'
     runtimeOnly 'org.apache.logging.log4j:log4j-slf4j18-impl:2.18.0'
 
+    implementation 'club.minnced:discord-webhooks:0.8.2'
+
     implementation 'org.jooq:jooq:3.17.2'
 
     implementation 'io.mikael:urlbuilder:2.0.9'
