--- application/src/main/java/org/togetherjava/tjbot/logging/discord/api/package-info.java
@@ -1,12 +1,0 @@
-/**
- * The Jackson representation of Discords Webhooks: <a href=
- * "https://discord.com/developers/docs/resources/webhook#execute-webhook-jsonform-params">API
- * Documentation</a>
- */
-@MethodsReturnNonnullByDefault
-@ParametersAreNonnullByDefault
-package org.togetherjava.tjbot.logging.discord.api;
-
-import org.togetherjava.tjbot.annotations.MethodsReturnNonnullByDefault;
-
-import javax.annotation.ParametersAreNonnullByDefault;
